chore(#5437): extract stream-expando subsystem to streams/expando.rs
The #5437 stream-expando store pushed streams.rs to 2008 lines, over the
2000-line file-size gate. Move the STREAM_EXPANDO table, set hook, getter, and
GC scan into a sibling module (behavior-preserving); streams.rs re-exports the
getter for streams/subclass.rs. streams.rs back to 1960 lines.
